app/requirements.txt
amqp==1.4.9
anyjson==0.3.3
aylien-apiclient==0.6.0
billiard==3.3.0.23
boto==2.43.0
celery==3.1.24
Django==1.10.2
elasticsearch==5.0.0
futures==3.0.5
httplib2==0.9.2
kombu==3.0.37
lxml==3.7.2
oauthlib==2.0.0
python-docx==0.8.6
pytz==2016.7
redis==2.10.5
requests==2.11.1
requests-aws4auth==0.9
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
urllib3==1.18